### 2024-03-14 #6489 @editorialbot
[PRE REVIEW]: scikit-eo: A Python package for Remote Sensing Data Analysis
**Submitting author:** @ytarazona (<a href="http://orcid.org/0000-0002-5208-1004">Yonatan Tarazona Coronel</a>)
**Repository:** https://github.com/yotarazona/scikit-eo
**Branch with paper.md** (empty if default branch): main
**Version:** v0.2.23
