+++ .github/workflows/pypi.yml
@@ -8,7 +8,7 @@ on: push
   #  types: [created]
 
 jobs:
-  deploy:
+  build:
 
     runs-on: ubuntu-latest
 
