+++ README.md
@@ -10,6 +10,9 @@
 [![Downloads](https://raw.githubusercontent.com/yotarazona/scikit-eo/main/docs/images/docs-passing-brightgreen.svg)]()
 [![tests](https://github.com/yotarazona/scikit-eo/actions/workflows/tests.yml/badge.svg)](https://github.com/yotarazona/scikit-eo/actions/workflows/tests.yml)
 
+HTML: <a href="https://joss.theoj.org/papers/46bccc5be81d7ea886e05807cfe6790c"><img src="https://joss.theoj.org/papers/46bccc5be81d7ea886e05807cfe6790c/status.svg"></a>
+Markdown: [![status](https://joss.theoj.org/papers/46bccc5be81d7ea886e05807cfe6790c/status.svg)](https://joss.theoj.org/papers/46bccc5be81d7ea886e05807cfe6790c)
+
 
 <img src="https://raw.githubusercontent.com/yotarazona/scikit-eo/main/docs/images/scikit-eo_logo.jpg" align="right" width="220"/>
 
