+++ contributors/contributors.md
@@ -1,11 +1,10 @@
 ## **Maintainers and contributors**
 
-**Scikit-eo** is maintained by a core group of Python enthusiasts working in various international universities such as Pontificia Universidad Católica del Perú (Peru), University of Coimbra (Portugal), Adam Mickiewicz University in Poznań (Poland) and The school of geography and sustainable development, University of St Andrews (UK).
+**Scikit-eo** is maintained by a core group of Python enthusiasts working in various international universities such as i) University of Coimbra, Department of Earth Sciences and CITEUC, Portugal, ii) The school of geography and sustainable development, University of St Andrews (UK), iii) Adam Mickiewicz University in Poznań (Poland) and iv) Pontificia Universidad Católica del Perú (Peru).
 
 - [Yonatan Tarazona](https://github.com/yotarazona)
 - [Fernando Benitez-Paez](https://github.com/mfbenitezp)
 - [Jakub Nowosad](https://github.com/Nowosad)
-- [David Montero Loaiza](https://github.com/davemlz)
 
 ## **Contributors and collaborators**
 
