+++ README.md
@@ -206,4 +206,4 @@ This package was created with [Cookiecutter](https://github.com/cookiecutter/coo
 
 ## Logo inspiration
 
-The logo on the package is inspired by fog oasis known as "lomas", thanks to **María Miyasiro** (*geolomera*).
+The logo on the package is inspired by fog oasis known as "lomas", thanks **María Miyasiro** (*geolomera*).
