+++ paper/paper.md
@@ -10,7 +10,7 @@ tags:
 authors:
   - name: Yonatan Tarazona
     orcid: 0000-0002-5208-1004
-    affiliation: "1,4"
+    affiliation: 1
   - name: Fernando Benitez-Paez
     affiliation: 2
     orcid: 0000-0002-9884-6471
