### 2024-04-29 #6692 @editorialbot
[REVIEW]: scikit-eo: A Python package for Remote Sensing Data Analysis
**Submitting author:** @ytarazona (<a href="http://orcid.org/0000-0002-5208-1004">Yonatan Tarazona Coronel</a>)
**Repository:** https://github.com/yotarazona/scikit-eo
**Branch with paper.md** (empty if default branch): main
**Version:** 0.2.31
