### 2024-04-29 #6692 @mikemahoney218
👋🏼 @ytarazona, @dbuscombe-usgs, @KBodolai: this is the review thread for the paper. Just about all of our communications should will happen here from now on. :smile:

```@editorialbot generate my checklist```
