### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
Please provide some instructions for installation of all dependencies, and running in virtual environment
Hi, nice work here. I'm reviewing for JOSS https://github.com/openjournals/joss-reviews/issues/6692

I'm a regular python user and I use a lot of different packages with various dependencies that aren't always compatible. I believe I may be your 'typical' end-user, in that regard. I think, therefore, that providing some instruction for isolating/containerizing `scikit-eo` for use would be beneficial. For example, conda provides an easy solution

`conda create -n scikiteo python=3.10`

then

`conda activate scikiteo`

then finally

`pip install scikeo`

I also noticed that the pip command did not install all required dependencies, which I had to install manually

* matplotlib
* geopandas
* seaborn
* earthpy
* statsmodels

Is there a reason why the pip package doesn't install these dependencies? This would be a major problem for a novice python user. At a minimum, all dependencies should be listed [here](https://yotarazona.github.io/scikit-eo/installation/). Ideally, you package them up for the user. I installed all dependency packages using conda, except `earthpy`
