### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
I am now running the provided test script, which errors because of new dependencies not installed by `pip install scikeo`. Namely

* tensorflow
* pytest
