### 2024-05-28 yotarazona/scikit-eo#6 @dbuscombe-usgs
test_scikeo.py returns an error (related to a lack of installation instructions)
ping https://github.com/openjournals/joss-reviews/issues/6692

This issue is related to https://github.com/yotarazona/scikit-eo/issues/5 because it arises due to a lack of instructions over installing the dependencies

After installing the extra dependencies required to run `test_scikeo.py`, I get the following error

```
Traceback (most recent call last):
  File "/home/marda/scikit-eo/tests/test_scikeo.py", line 10, in <module>
    from scikeo.process import confintervalML
  File "/home/marda/miniconda3/envs/scikiteo/lib/python3.10/site-packages/scikeo/process.py", line 5, in <module>
    import fiona
  File "/home/marda/miniconda3/envs/scikiteo/lib/python3.10/site-packages/fiona/__init__.py", line 41, in <module>
    from fiona._env import (
ImportError: libminizip.so: cannot open shared object file: No such file or directory
```

The issue seems to be occurring when fiona is imported. It appears fiona was installed twice (once from pip and once from conda). I removed the conda version, and I could import fiona. However, that removed tensorflow and geopandas, too. I reinstalled tensorflow and the test script started to run, but then I got a (very strange) message

```
2024-05-28 10:10:31.849532: I tensorflow/core/util/port.cc:113] oneDNN custom operations are on. You may see slightly different numerical results due to floating-point round-off errors from different computation orders. To turn them off, set the environment variable `TF_ENABLE_ONEDNN_OPTS=0`.
2024-05-28 10:10:31.850110: I external/local_tsl/tsl/cuda/cudart_stub.cc:32] Could not find cuda drivers on your machine, GPU will not be used.
2024-05-28 10:10:31.852704: I external/local_tsl/tsl/cuda/cudart_stub.cc:32] Could not find cuda drivers on your machine, GPU will not be used.
2024-05-28 10:10:31.886452: I tensorflow/core/platform/cpu_feature_guard.cc:210] This TensorFlow binary is optimized to use available CPU instructions in performance-critical operations.
To enable the following instructions: AVX2 AVX512F AVX512_VNNI FMA, in other operations, rebuild TensorFlow with the appropriate compiler flags.
2024-05-28 10:10:32.566326: W tensorflow/compiler/tf2tensorrt/utils/py_utils.cc:38] TF-TRT Warning: Could not find TensorRT
Traceback (most recent call last):
  File "/home/marda/scikit-eo/tests/test_scikeo.py", line 21, in <module>
    from scikeo.deeplearning import DL
  File "/home/marda/miniconda3/envs/scikiteo/lib/python3.10/site-packages/scikeo/deeplearning.py", line 7, in <module>
    from tensorflow.keras import models
ModuleNotFoundError: No module named 'tensorflow.keras'

```

It appears you need to provide the user with a computing environment that installs all of the dependencies. This isn't a trivial detail, but a requirement.
