### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
This is related to https://github.com/yotarazona/scikit-eo/issues/6
