### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
It would appear jupyter is also missing as a dependency in the pip package, which is required to run all of the notebooks in the `example` folder. And `google` is required to access example datasets
