### 2024-05-28 yotarazona/scikit-eo#7 @dbuscombe-usgs
Notebooks don't run locally without modifcation. Guidance would be helpful
ping JOSS https://github.com/openjournals/joss-reviews/issues/6692

I am trying to run the provided notebooks from source, not on colab. I shouldn't need to run the examples on any third-party platform like colab. They should run locally

None of the notebooks in the examples folder run, because

1. jupyter and google are not found and need to be installed separately. This would need to be communicated to the user, see https://github.com/yotarazona/scikit-eo/issues/5
2. the datasets are not accessible, for example "/content/drive/MyDrive/Packages/scikit-eo_data/01_machine_learning/LC08_232066_20190727_SR.tif" doesn't exist
