### 2024-05-28 yotarazona/scikit-eo#7 @dbuscombe-usgs
I was able to find the dataset for the first notebook, 'Machine Learning' on the webiste [here](https://yotarazona.github.io/scikit-eo/get-started/). Thanks for providing these data. I was able to comment out the google drive cells, update the paths to my locally downloaded data, and run the first notebook.

Perhaps just some more clarification should be provided, namely
1. that it is possible to run the notebooks locally, with some modifications and by downloading the data locally
2. also providing a stable computational environment will all needed dependencies
