### 2024-05-28 yotarazona/scikit-eo#8 @dbuscombe-usgs
Error in notebook 2. Unknown variable `n`
ping https://github.com/openjournals/joss-reviews/issues/6692

The following line
```result = confintervalML(matrix = values, image_pred = img, pixel_size = 30, nodata = -9999)```

returns the following error

```
UnboundLocalError                         Traceback (most recent call last)
Cell In[18], line 1
----> 1 result = confintervalML(matrix = values, image_pred = img, pixel_size = 30, nodata = -9999)

File ~/miniconda3/envs/scikiteo/lib/python3.10/site-packages/scikeo/process.py:251, in confintervalML(matrix, image_pred, pixel_size, conf, nodata)
    248 # copy of matrix
    249 matConf = matrix.astype(float).copy()
--> 251 for i in range(n):
    252     matConf[i,:] = (matConf[i,:]/ni[i])*wi[i]
    254 # user's accuracy

UnboundLocalError: local variable 'n' referenced before assignment
```

this seems to be an error related to a variable, `n`, not being specified in the scikit-eo function `confintervalML`
