### 2024-05-28 yotarazona/scikit-eo#7 @dbuscombe-usgs
I note that you have done a good job at linking to datasets in other notebooks other than number 1. I think the only things required are
1. a link in the first notebook to the data
2. a higher-level description of how to run locally
