### 2024-05-28 yotarazona/scikit-eo#9 @dbuscombe-usgs
Error in notebook 10_Logistic_regression_in_remote_sensing.ipynb
ping https://github.com/openjournals/joss-reviews/issues/6692

Heads-up that the link provided does not contain the data `Variables.tif`, which is what is required to run the example. Please update with the new link to the data
