### 2024-05-28 yotarazona/scikit-eo#10 @dbuscombe-usgs
Tensorflow import error in notebook 11
ping https://github.com/openjournals/joss-reviews/issues/6692

Tensorflow is not a listed dependency for your package. See https://github.com/yotarazona/scikit-eo/issues/6. This means that notebook 11 fails

I have tensorflow installed, and can verify using python and ipython terminals, as well as jupyter. I can import `import tensorflow as tf` with no error

However, when I attempt `from scikeo.deeplearning import DL`, I get the following error `ModuleNotFoundError: No module named 'tensorflow.keras'`

It appears the packaging issues I've already identified are at play here (i.e., these are all parts of the same problem). In this case, it seems to me that the from `scikeo.deeplearning` module is improperly linked against tensorflow, or perhaps the conda environment I had to create to run these codes was not done so optimally? For example, I regularly have to install tensorflow in some specific order.

The packages I maintained usually spend quite a significant amount of time ensuring that all the dependencies are installable together in a stable way, usually tested on multiple machines. I recommend that here, coming up with a conda or virtual environment recipe that is known to work for all the provided functionality of this software.
