### 2024-05-28 yotarazona/scikit-eo#11 @ddbuscombe
Error in notebook 14, writeRaster with outputs
ping https://github.com/openjournals/joss-reviews/issues/6692

Heads-up that the last cell of notebook 14 contains an error. The cell is

```
writeRaster(arr = clip_image, image = img, filename = 'LC08_band5_crop',
   filepath = '/content/drive/MyDrive/Packages/scikit-eo_data/01_machine_learning', n = 1)
```

However, the variable `img` has not been set. If I use the variable `clip_image`, defined in the previous cells, I get an new error

```
Cell In[12], line 1
----> 1 writeRaster(arr = clip_image, image = clip_image, filename = 'LC08_band5_crop',
      2    filepath = '/content/drive/MyDrive/Packages/scikit-eo_data/01_machine_learning', n = 1)

File ~/miniconda3/envs/scikiteo/lib/python3.10/site-packages/scikeo/writeRaster.py:31, in writeRaster(arr, image, filename, filepath, n)
     28     raise TypeError(f'{arr} must be an array with 2d (rows and cols).')
     30 if not isinstance(image, (rasterio.io.DatasetReader)):
---> 31     raise TypeError(f'The image must be rasterio.io.DatasetReader. image = {type(image)}')
     33 if not arr.ndim >= 2:
     34     raise ValueError(f'Input mus have two or three dimensions. ndim = {arr.ndim}')

TypeError: The image must be rasterio.io.DatasetReader. image = <class 'numpy.ndarray'>
```

I think you need to add one more line of code, that creates `img` as a `rasterio.io.DatasetReader.image` type, converted from numpy array `clip_image`?
