### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
I have been trying to troubleshoot the issues I am having with tensorflow (see also https://github.com/yotarazona/scikit-eo/issues/10) and, more specifically, `from scikeo.deeplearning import DL`, which returns the following error

```
ModuleNotFoundError: No module named 'tensorflow.keras'
```

This is despite being able to `import tensorflow`. Very bizarre. I realize this may be a tensorflow issue, not a `scikit-eo` issue, but I'm still confused as to why you haven't listed tensorflow as a dependency in https://github.com/yotarazona/scikit-eo/blob/main/requirements.txt or anywhere else. Have you at least tested what versions are known to be stable? Any other guidance you could provide? Thanks
