### 2024-05-28 yotarazona/scikit-eo#5 @dbuscombe-usgs
I managed to fix this by modifying `deeplearning.py`, in which I replaced

```
from tensorflow.keras import models
from tensorflow.keras import layers
from tensorflow.keras.utils import to_categorical
```

with

`from tensorflow import keras`

then made the necessary edits to calls to `layers` and `models`. Then reinstalled locally using `pip install -e .`

Now I can `from scikeo.deeplearning import DL` without error
