### 2024-05-28 yotarazona/scikit-eo#6 @dbuscombe-usgs
After some troubleshooting, I was able to fix the tensorflow error above by modifying `scikeo/deeplearning.py` - please see https://github.com/yotarazona/scikit-eo/issues/5#issuecomment-2136179777 for more details. I would recommend implementing this change
