### 2024-05-28 yotarazona/scikit-eo#12 @dbuscombe-usgs
Bug in notebook 11: type error (shapefile)
ping https://github.com/openjournals/joss-reviews/issues/6692

Cell 7 in notebook 11
```
endm = extract(img, endm)
```

returns the following error

```
Cell In[7], line 2
      1 # endmembers
----> 2 endm = extract(img, endm)
      3 endm

File ~/scikit-eo/scikeo/process.py:136, in extract(image, shp)
    133     raise TypeError('"image" must be raster read by rasterio.open().')
    135 if not isinstance(shp, (gpd.geodataframe.GeoDataFrame)):
--> 136     raise TypeError('"shp" must be Shapefile (.shp)')
    138 gdf = gpd.GeoDataFrame(geometry = shp['geometry'])
    140 coords = [(x,y) for x, y in zip(gdf['geometry'].x, gdf['geometry'].y)]

TypeError: "shp" must be Shapefile (.shp)
```

The variables above have been defined like so, using the files downloaded from your google drive links
```
path_raster = r'/home/marda/Downloads/11_deepLearning/LC08_232066_20190727_SR.tif'
img = rasterio.open(path_raster)

path_endm = r"/home/marda/Downloads/11_deepLearning/endmembers.dbf"
endm = DBF(path_endm)
```
