### 2024-05-28 yotarazona/scikit-eo#10 @dbuscombe-usgs
I was able to fix this here https://github.com/yotarazona/scikit-eo/issues/5#issuecomment-2136179777

I would suggest this way of importing keras to avoid those import errors
