### 2024-05-28 yotarazona/scikit-eo#13 @dbuscombe-usgs
JOSS paper: please briefly describe how this software compares to other commonly-used packages
Hi, I'm really enjoying reviewing your codes and paper for JOSS https://github.com/openjournals/joss-reviews/issues/6692

As you will have seen in previous issues, I have found a few minor issues related to imports, data links, one or two hopefully easily fixable bugs. I have however managed to run almost all of the notebooks, read the paper, review and test API functionality, and I'm very pleased with this new addition to the community!

In order for me to sign off on the paper, I (and I assume many readers) would like to know the community a little better. Specifically, with reference to review criterion

```State of the field: Do the authors describe how this software compares to other commonly-used packages?```

I note that in the paper there are no references to other software packages that may have overlapping functionality, or perhaps legacy software that this package now replaces? Please provide some overview of this in a revised manuscript. For example, I am a user of geemap, which I would consider to be comparable to this package. A very small literature review of what else has been done in support of community earth observation python software would be very helpful. Thanks.
