### 2024-05-28 yotarazona/scikit-eo#14 @dbuscombe-usgs
None of the links on the API webpages work
ping https://github.com/openjournals/joss-reviews/issues/6692

Every single 'SOURCE' link on the API pages of the website is broken (404 - Not found)
