### 2024-05-28 yotarazona/scikit-eo#15 @dbuscombe-usgs
JOSS paper: please describe the deep learning models
ping https://github.com/openjournals/joss-reviews/issues/6692

I think it is great that you have provided a deep learning model for landcover classification, thanks for that. Notwithstanding the fact that I haven't yet been able to run the provided notebook (due to https://github.com/yotarazona/scikit-eo/issues/12), I was wondering if you could provide more details about the model, the data it is trained with (it is trained?), the classes it is designed to predict, and other important details, like what data bands it uses for prediction, and how it determines class (softmax?).

There is no mention of those details on the website https://yotarazona.github.io/scikit-eo/deeplearning/ nor the paper. It would appear this is a model that is trained on the fly, rather than pretrained, with weights downloaded from somewhere? Am I right? Also, what makes it 'deep' learning as opposed to machine learning. Those details would be great in the paper and on the website
