### 2024-05-28 yotarazona/scikit-eo#13 @dbuscombe-usgs
On the same thread, it would be helpful to provide an overview of what need the 'deep learning' module meets, in line with more details about what it is and how it works https://github.com/yotarazona/scikit-eo/issues/15
