### 2024-05-28 #6692 @dbuscombe-usgs
> Hi @dbuscombe-usgs and @KBodolai ! I wanted to bump this now that we're about a month into the review window, and ask how your reviews were progressing/if you're still expecting to complete your reviews on the original timeline.
>
> Just as a reminder, the first step in the review is to post `@editorialbot generate my checklist` as the start of a new comment in this thread, which will generate a review checklist for you to use!

Sorry for the delay, and thanks for the reminder. I was able to devote all day to this today, reading and testing everything. This is *mostly* great and *mostly* works as described, but I have a few issues I'd like to see tackled, and one requested edit for the paper. I'd be an end-user of this software for sure, and I'm sure it will be valuable for the community. I'll be in touch soon!
