### 2024-06-02 e462206
installing scikit-eo
