+++ setup.py
@@ -52,6 +52,6 @@ setup(
     test_suite='tests',
     tests_require=test_requirements,
     url='https://github.com/yotarazona/scikit-eo',
-    version='0.2.23',
+    version='0.2.24',
     zip_safe=False,
 )
