### 2024-06-02 yotarazona/scikit-eo#9 @yotarazona
Hi @dbuscombe-usgs, thanks for your time and your review of the package.
Done, all links were updated.
