### 2024-06-02 yotarazona/scikit-eo#12 @yotarazona
Hi @dbuscombe-usgs, thanks!. The shapefile used was updated, now

`path_raster = r"C:\Users\USER\Desktop\Github_Repo\scikiteo_data\11_deepLearning\LC08_232066_20190727_SR.tif"
img = rasterio.open(path_raster)

path_endm = r"C:\Users\USER\Desktop\Github_Repo\scikiteo_data\11_deepLearning\endmembers.shp"
endm = gpd.read_file(path_endm)
`
