### 2024-06-02 yotarazona/scikit-eo#11 @yotarazona
Hi @dbuscombe-usgs, Done!.
A line with the "img" variable was missing:

```python
original_image = rasterio.open(path_raster).read()
original_image = np.moveaxis(original_image, 0, -1)

img = rasterio.open(output_path_raster + '/' + output_name + '.tif')
clip_image = img.read()
clip_image = np.moveaxis(clip_image, 0, -1)
```
