### 2024-06-04 yotarazona/scikit-eo#8 @yotarazona
Hi @dbuscombe-usgs, thanks for your feedback. This notebook was run but I didn't get any error related to the variable ```n```. I am using the v0.2.23. Tell me if this error still persists.
