### 2024-06-04 yotarazona/scikit-eo#7 @yotarazona
Thank you so much @dbuscombe-usgs !. Done, each notebook has a correct link to download data files. A higher-level description of how to run data locally was added in the new version of the package.
