### 2024-06-04 yotarazona/scikit-eo#5 @yotarazona
Thank you so much @dbuscombe-usgs !. All required dependencies was add: https://github.com/yotarazona/scikit-eo/blob/main/requirements.txt. By installing ```scikit-eo``` all these dependencies will be installed as well in this new version.  This dependencies used by  ```scikit-eo``` was explain in the README.md (https://github.com/yotarazona/scikit-eo). Additionally, isolating/containerizing ```scikit-eo``` was also added in the README.md as a option. This will clarify all required dependencies.
Many novice python users don't know about creating environments, so I left containerization as an additional option.

Regarding the class ```deeplearning.py```, it was fix.

```python
from tensorflow import keras
from keras import models
from keras import layers
from keras.utils import to_categorical
```
