### 2024-06-04 yotarazona/scikit-eo#6 @yotarazona
Thank you so much @dbuscombe-usgs !. You are right, all required dependencies will be installed by installing ```scikit-eo```: https://github.com/yotarazona/scikit-eo/blob/main/requirements.txt. An explanation was added in the README.md file.
#5
