### 2024-06-04 8912fc6
Update deeplearning.py
