### 2024-06-05 yotarazona/scikit-eo#16 @KBodolai
Nit: Creating an issue has non-working options
Hi, starting some issues related to the JOSS Review openjournals/joss-reviews/issues/6692:

When creating a new issue, the following options are given:
![image](https://github.com/yotarazona/scikit-eo/assets/69843715/9aa4cfe7-ee25-406e-931a-1db887282306)

But the `Ask questions` and `Ideas` options return a 404 error. This could be easily fixed by adding templates for it.
