### 2024-06-10 bd55108
updates gha templates
