+++ .github/ISSUE_TEMPLATE/questions.md
@@ -1,0 +1,18 @@
+---
+name: Ask questions
+about: Ask a question about the package
+labels: questions
+---
+
+<!-- Please search existing issues to avoid creating duplicates. -->
+
+### Description
+
+Ask a question.
+If you have a code example, also add it.
+
+### Code example
+
+```
+Paste the command(s) you ran and the output.
+```
