+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,10 +1,4 @@
 contact_links:
-    - name: Ask questions
-      url: https://github.com/yotarazona/scikit-eo/discussions/categories/q-a
-      about: Please ask and answer questions here.
-    - name: Ideas
-      url: https://github.com/yotarazona/scikit-eo/discussions/categories/ideas
-      about: Please share your ideas here.
     - name: Ask questions from the GIS community
       url: https://gis.stackexchange.com
       about: To get answers from questions in the GIS commminuty, please ask and answer questions here.
