### 2024-06-10 yotarazona/scikit-eo#16 @yotarazona
Hi @KBodolai, thanks for your suggestion. Done! a template was added.
