### 2024-06-10 yotarazona/scikit-eo#10 @yotarazona
Hi @dbuscombe-usgs, thanks a lot for open this issue!. After some trial and error the package is running well now. ```tensorflow```, ```rasterio``` and other required dependencies will be installed when ```scikit-eo``` is installed. Additionally, ```scikit-eo``` was tested on multiple machines in order to make sure everything is working out. Please see issue number #5.
