### 2024-06-10 #6692 @yotarazona
Hi @mikemahoney218, @dbuscombe-usgs and @KBodolai, thanks a lot for your effort and time spent reviewing this python package. Very grateful for that!. There are only two missing issues regarding the paper that we will finish during the following days. Let us know if any additional issue needs to be addressed.
