### 2024-06-10 yotarazona/scikit-eo#14 @dbuscombe-usgs
All of the 'SOURCE' links are broken

I just checked and they are all still down, e.g.

https://yotarazona.github.io/scikit-eo/scikeo/pca.py#L0
