+++ paper/paper.md
@@ -67,7 +67,14 @@ In this paper, we introduce **scikit-eo**, a specialized library with tailored a
 
 In environmental studies, particularly in land cover classification, the availability of scalable but user-friendly software tools is crucial for facilitating research, analysis, and modelling. With the widespread adoption of Python as one of the most popular programming languages, particularly in the GIScience and remote sensing fields, developing specialized packages has massively enhanced the effectiveness of environmental research. **scikit-eo** is a dedicated piece of software tailored to address unique challenges encountered in land cover mapping, forest or land degradation [@TARAZONA2020100337], and the fusion of multiple satellite imagery from several formats [@Tarazona2021]. **scikit-eo** provides the assessment and calibration metrics to evaluate the provided outputs. The current version **scikit-eo** requires that users provide the dataset to process. However, we expect to provide a wide range of functionalities for acquiring environmental data from diverse sources and file formats, enabling researchers to access satellite imagery.
 
-One of **scikit-eo's** key strengths is its advanced analysis capabilities. It provides a rich suite of algorithms specifically designed for environmental studies, including statistical analysis, machine learning, deep learning, data fusion, and spatial analysis. Researchers can leverage these tools to explore patterns, relationships, and trends within their datasets, uncover complex land or forest degradation or mapping, classify the land cover, and generate insightful visualizations.
+One of **scikit-eo's** key strengths is its advanced analysis capabilities. It provides a rich suite of algorithms specifically designed for environmental studies, including statistical analysis, deep learning, data fusion, and spatial analysis. Researchers can leverage these tools to explore patterns, relationships, and trends within their datasets, uncover complex land or forest degradation or mapping, classify the land cover, and generate insightful visualizations.
+
+As a particular example of these advanced analysis capabilities, we have integrated the **`deepLearning`** function, which includes the *Fully connected layers (FC)*, also known as *dense layers* model. This is one of the most straightforward yet functional neural networks we can apply to remote sensing analysis. The term "fully connected" comes from the fact that each neuron in one layer is connected to every neuron in the preceding layer, creating a highly interconnected network known as the Multi-Layer Perceptron (MLP). The model is trained using a specified dataset, with the bands as *input_shape*, including input features and corresponding class labels. The weights 𝑊 are initialized and then adjusted during training to ensure the neural network's output is consistent with the class labels. The training process involves minimizing the error function using Gradient Descent and backpropagation algorithms. The activation functions used in this model are ReLU (*Rectified Linear Unit*) for the neurons in each hidden layer and *Softmax* for the final classification layer. 
+
+It's important to note that *ReLU* introduces non-linearity to the model, enabling it to learn complex patterns, while *Softmax*is used for multi-class classification, transforming the output into a probability distribution over multiple classes and better suited for more than two land covers. Unlike traditional machine learning models, such as linear regression or decision trees, which typically don't involve multiple layers of abstraction, the FC model uses multiple hidden layers, allowing it to learn hierarchical representations of the input data. Deep learning models like this one can automatically learn and extract complex features from the raw input, making them exceptionally powerful for tasks such as land cover classification.
+
+To run an example of how to use the function **`deepLearning`**  find a detailed notebook in tutorial No 11 [Deep Learning Classification](https://github.com/yotarazona/scikit-eo/blob/main/examples/11_Deep_Learning_Classification_FullyConnected.ipynb).
+
 
 ### **scikit-eo** in the lecture room:
 
