+++ paper/paper.md
@@ -71,7 +71,7 @@ One of **scikit-eo's** key strengths is its advanced analysis capabilities. It p
 
 As a particular example of these advanced analysis capabilities, we have integrated the **`deepLearning`** function, which includes the *Fully connected layers (FC)*, also known as *dense layers* model. This is one of the most straightforward yet functional neural networks we can apply to remote sensing analysis. The term "fully connected" comes from the fact that each neuron in one layer is connected to every neuron in the preceding layer, creating a highly interconnected network known as the Multi-Layer Perceptron (MLP). The model is trained using a specified dataset, with the bands as *input_shape*, including input features and corresponding class labels. The weights 𝑊 are initialized and then adjusted during training to ensure the neural network's output is consistent with the class labels. The training process involves minimizing the error function using Gradient Descent and backpropagation algorithms. The activation functions used in this model are ReLU (*Rectified Linear Unit*) for the neurons in each hidden layer and *Softmax* for the final classification layer. 
 
-It's important to note that *ReLU* introduces non-linearity to the model, enabling it to learn complex patterns, while *Softmax*is used for multi-class classification, transforming the output into a probability distribution over multiple classes and better suited for more than two land covers. Unlike traditional machine learning models, such as linear regression or decision trees, which typically don't involve multiple layers of abstraction, the FC model uses multiple hidden layers, allowing it to learn hierarchical representations of the input data. Deep learning models like this one can automatically learn and extract complex features from the raw input, making them exceptionally powerful for tasks such as land cover classification.
+It's important to note that *ReLU* introduces non-linearity to the model, enabling it to learn complex patterns, while *Softmax*is used for multi-class classification, transforming the output into a probability distribution over multiple classes and better suited for more than two land covers. Unlike traditional machine learning models, such as linear regression or decision trees, which typically don't involve multiple layers of abstraction, the FC model uses multiple hidden layers, allowing it to learn hierarchical representations of the input data. Deep learning models like this one can automatically learn and extract complex features from the raw input, making them exceptionally powerful for tasks such as land cover classification. 
 
 To run an example of how to use the function **`deepLearning`**  find a detailed notebook in tutorial No 11 [Deep Learning Classification](https://github.com/yotarazona/scikit-eo/blob/main/examples/11_Deep_Learning_Classification_FullyConnected.ipynb).
 
@@ -108,6 +108,10 @@ As open-source software keeps transforming the landscape of scientific research 
 : Main tools available for **scikit-eo** package. \label{table:1}
 
 For more information the reader is referred to the [scikit-eo](https://yotarazona.github.io/scikit-eo/) website.
+
+# State of the field:
+
+**Scikit-eo** is built upon well-known packages and libraries that support remote sensing analysis in python, but is designed specifically to study land cover classification, and providing tailored functionalities for environmental studies. It aims to simplify the identification of patterns, changes, and trends for environmental research. Like many other Python packages, **Scikit-eo**** makes use of *Rasterio* and *GDAL*, which are essential for geospatial data handling. Additionally, **Scikit-eo** is based on **Scikit-learn**, a widely-used machine learning library in Python that includes several algorithms for classification, regression, clustering, and dimensionality reduction. However, **Scikit-eo** includes machine learning tools to remote sensing analysis. Another well-known library, *Geemap* that provides interactive mapping capabilities using Google Earth Engine and focuses more on visualization and data exploration, while **Scikit-eo** is designed for deeper analytical tasks using remote sensing data for land cover analysis. Lastly, *EO-learn* can be considered the most aligned package with **Scikit-eo** providing efficient processing and analysis of satellite imagery capabilities but lacking the deep learning and tailored functions specifically included for land cover analysis.
 
 # Acknowledgments
 
