### 2024-06-17 yotarazona/scikit-eo#15 @yotarazona
Hi @dbuscombe-usgs, sorry for the delay!. This issue was cover in the new version of the paper: https://github.com/yotarazona/scikit-eo/actions/runs/9550361620. Details of the Deep Learning model was added in the notebook tutorial number [11](https://github.com/yotarazona/scikit-eo/blob/main/examples/11_Deep_Learning_Classification_FullyConnected.ipynb).

Thanks a lot for your suggestions which was used to improve ```scikit-eo```.
