### 2024-06-17 yotarazona/scikit-eo#13 @yotarazona
Hi @dbuscombe-usgs, sorry for the delay!.
Thanks for this!. This issue was cover in the new version of the paper as well. Please see https://github.com/yotarazona/scikit-eo/actions/runs/9550361620 for more details. In addition, details of the Deep Learning model including activation function used, among other details were added in the notebook tutorial number [11](https://github.com/yotarazona/scikit-eo/blob/main/examples/11_Deep_Learning_Classification_FullyConnected.ipynb).

We hope we have met all your expectations.
