### 2024-06-17 yotarazona/scikit-eo#14 @yotarazona
Hi @dbuscombe-usgs. Please find attached a gif where you can see all the links related to the package are working.

![gif](https://github.com/yotarazona/scikit-eo/assets/143953867/b470b293-1435-47d3-9d38-4ac09dcb3f8d)

Maybe you have some old version of the website in your browser cache.
Best.
