### 2024-06-17 yotarazona/scikit-eo#8 @yotarazona
Did you make sure if this issue is still doesn´t work out?. @dbuscombe-usgs
