### 2024-06-17 yotarazona/scikit-eo#14 @dbuscombe-usgs
![image](https://github.com/yotarazona/scikit-eo/assets/3596509/65a2ac9f-9ac8-48d7-970d-446d60decd10)

Thanks for the gif. I'm only talking about the SOURCE links on the API pages. It's not a big deal, but probably an easy fix. My cache is clear. This is a real bug.
