### 2024-06-17 yotarazona/scikit-eo#8 @dbuscombe-usgs
I'll review later today
