### 2024-06-17 yotarazona/scikit-eo#15 @dbuscombe-usgs
Thanks
