### 2024-06-17 yotarazona/scikit-eo#8 @yotarazona
Great! Thanks
