### 2024-06-17 yotarazona/scikit-eo#8 @dbuscombe-usgs
Look good. I can sign off on this now. Thanks for addressing all these comments!
