### 2024-06-17 #6692 @dbuscombe-usgs
I have completed my review of the manuscript and software. Thanks to @yotarazona and team for being responsive. All my previous issues have been satisfactorily resolved (except a minor one involving updating website links). I have completed my checklist above. Details have been added about the deep learning model and links to previous research. Some bugs have been fixed, and clearer instructions are provided for new users. So, I am happy to sign off on my review. I think this will become a valuable contribution to the remote sensing software community for both teaching and research.
