### 2024-06-17 #6692 @yotarazona
Thank you very much @dbuscombe-usgs for your time and effort to review the package. I am sure that step by step the python package will be used by the remote sensing community, that is the idea. I am glad to have contributed with this small package (```scikit-eo```).
Many thanks!.
