### 2024-06-18 yotarazona/scikit-eo#16 @KBodolai
fantastic, thanks!
