### 2024-06-18 yotarazona/scikit-eo#10 @KBodolai
Hi @yotarazona , I'm moving on with my side of the review, and I have a question about this.

Right now it installs everything, which is the most straightforward way, however, tensorflow makes the installation quite heavy.

I've seen in the past packages that have the possibility of using neural networks but where it's not a strict requirement treat it as an optional dependency, that can be installed if you do something like `pip install scikeo[neural-nets]`.

I think it could improve the user experience, as long as it's documented in the installation instructions, although it depends a lot on what you expect the average user to be doing!
