### 2024-06-18 yotarazona/scikit-eo#18 @KBodolai
Extracting endmembers (sampling points from raster) expected behaviour
ping openjournals/joss-reviews#6692

### Description

I was stress-testing the examples thinking of common mistakes an inexperienced user may make. In the first machine learning example, if you have some point endmembers with a different CRS, it returns, without any warning or complaint, a  dataframe where all the values are -9999.

This is not an issue of the example per-se, but I'm thinking it could be useful for `scikeo.process.extract` to warn the user an attempt to reproject (or just warn them that when the CRS is not the same the results will likely be wrong).

What do you think @yotarazona ?
