### 2024-06-18 yotarazona/scikit-eo#19 @KBodolai
Examples 
ping openjournals/joss-reviews#6692

First of all, I want to congratulate you on the structure and how the notebooks walk the user through using ML with remote sensing data, I think there's some pretty powerful abstractions here. I'll be pointing out some minor issues or things to improve here and in some other issues for the joss review, but I just wanted to clarify that I think this is great work!

There's a few things in the notebooks that I think can be improved:

1. The 'running notebooks' section (cell 8) seems to be common for all notebooks. Since currently the notebooks have a default that will work wherever you run it, I think it would be better for this to form part of a readme in the notebooks folder, and perhaps use that space to add some more information / context in the notebooks.
2. More verbosity in examples - I think that a few of the examples could do with a bit more text explaining what's going on, where the data comes from, contextualising the results, key scientific work supporting the methods...
3. Adding information about running things at scale - a big challenge of using ML with remote sensing is the size of our data once we try to run things for e.g. entire countries. How can we scale things up? Can we add an example showing how to use modern software such as dask + xarray to scale?
4. Where does the endmember matrix come from in Example 06?
5. IN the data fusion examples, the images are already well aligned and processed (I'm guessing the S1 data has been downscaled to 30m and the grids have been aligned). This is a strong assumption for working with satellite data and I think it should be explained (which is in the documentation of the `fusionrs` function, kudos!), but I think it wouldn't hurt to add it in the notebook as well (particularly thinking about the use of the software for teaching!)
