### 2024-06-18 yotarazona/scikit-eo#20 @KBodolai
Misleading documentation on `extract` function in `scikeo.process`
ping openjournals/joss-reviews#6692

The documentation for the `extract` function states that it extracts raster values using a shapefile.

According to the docs it says it takes an `image` (rasterio.io.DatasetReader) and a `shp` (Vector file, typically shapefile.). However, the function only works with point geometries, and not general vector data. I think the function could be modified to work with more geometries, like polygons, but at the very least the fact that it only works with points should be in the documentation, and probably there should be a check in the code that alerts the user.

The links to the source code in the docs are also broken,  when I click the `source` button in the top right of the screenshot below, it returns a 404 error.

![image](https://github.com/yotarazona/scikit-eo/assets/69843715/64fec781-7026-403a-b6ee-2fd5b066d48d)

Additionally, as a nit, I'd change the description to say 'vector file' instead of shapefile, since it is more general than that.
