### 2024-06-18 yotarazona/scikit-eo#14 @KBodolai
I'm getting the same as @dbuscombe-usgs above. I think it's because the links it points to are something like:

`https://yotarazona.github.io/scikit-eo/scikeo/mla.py#L23`

where it should be:

`https://github.com/yotarazona/scikit-eo/blob/main/scikeo/mla.py#L23`
