### 2024-06-21 27694d6
Update atmosCorr.md
