+++ docs/atmosCorr.md
@@ -1,6 +1,6 @@
 <!-- markdownlint-disable -->
 
-<a href="..\scikeo\atmosCorr.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L0"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 # <kbd>module</kbd> `atmosCorr`
 
@@ -11,12 +11,12 @@
 
 ---
 
-<a href="..\scikeo\atmosCorr.py#L8"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L8"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ## <kbd>class</kbd> `atmosCorr`
 Atmospheric Correction in Optical domain 
 
-<a href="..\scikeo\atmosCorr.py#L12"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L12"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ### <kbd>method</kbd> `__init__`
 
@@ -35,7 +35,7 @@ Parameter:
 
 ---
 
-<a href="..\scikeo\atmosCorr.py#L203"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L203"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ### <kbd>method</kbd> `DOS`
 
@@ -64,7 +64,7 @@ Chavez, P.S. (1988). An Improved Dark-Object Subtraction Technique for Atmospher
 
 ---
 
-<a href="..\scikeo\atmosCorr.py#L48"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L48"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ### <kbd>method</kbd> `RAD`
 
@@ -90,7 +90,7 @@ Return:  An array with radiance values with 3d, i.e. (rows, cols, bands).
 
 ---
 
-<a href="..\scikeo\atmosCorr.py#L126"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
+<a href="https://github.com/yotarazona/scikit-eo/blob/main/scikeo/atmosCorr.py#L126"><img align="right" style="float:right;" src="https://img.shields.io/badge/-source-cccccc?style=flat-square"></a>
 
 ### <kbd>method</kbd> `TOA`
 
