### 2024-06-21 fadb14b
Update deeplearning.md
