+++ scikeo/__init__.py
@@ -1,5 +1,7 @@
 """Top-level package for scikit-eo."""
 
+import importlib.metadata
+
 __author__ = """Yonatan Tarazona Coronel"""
 __email__ = 'geoyons@gmail.com'
-__version__ = '0.1.0'
+__version__ = importlib.metadata.version('scikeo')
