### 2024-06-21 f81f1e5
improving geometry type by extracting values from a raster image
