### 2024-06-22 yotarazona/scikit-eo#20 @yotarazona
Hi @KBodolai, thank you for this feedback!. In the new version of the package a geometry type was add inside the ```extract()```.  The function ```extract()``` now alerts the user if the input vector is of a type different than point. In addition, the function also alerts if the image and vector projection systems are different, please see https://github.com/yotarazona/scikit-eo/blob/main/scikeo/process.py#L112.

In a future version, a polygon geometry will be added, don't doubt that.
