### 2024-06-22 yotarazona/scikit-eo#14 @yotarazona
Hi @KBodolai, thanks for this unnoticed tip!. All links were updated now :).
