### 2024-06-22 yotarazona/scikit-eo#18 @yotarazona
Hi @KBodolai, thanks again for this feedback. It was fixed in the new version of the package. #20
