+++ docs/tutorials.md
@@ -13,39 +13,39 @@ Remote sensing tools of **scikit-eo** can be mainly grouped into 3 sets.
 
 ### Tools for remote sensing data analysis
 
-1 [Machine Learning.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/01.%20Machine_Learning.ipynb)
+1 [Machine Learning.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/01.%20Machine_Learning.ipynb)
 
-2 [Estimated area and uncertainty in Machine Learning.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/02.%20Estimated%20area%20and%20uncertainty%20in%20Machine%20Learning.ipynb)
+2 [Estimated area and uncertainty in Machine Learning.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/02.%20Estimated%20area%20and%20uncertainty%20in%20Machine%20Learning.ipynb)
 
-3 [Calibrating supervised classification in Remote Sensing.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/03_Calibrating_supervised_classification_in_Remote_Sensing.ipynb)
+3 [Calibrating supervised classification in Remote Sensing.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/03_Calibrating_supervised_classification_in_Remote_Sensing.ipynb)
 
-4 [Kmeans classification.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/04_K_means_classification.ipynb)
+4 [Kmeans classification.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/04_K_means_classification.ipynb)
 
-5 [Fusion of radar and optical images.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/05_Fusion_of_radar_and_optical_images.ipynb)
+5 [Fusion of radar and optical images.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/05_Fusion_of_radar_and_optical_images.ipynb)
 
-6 [Spectral Mixture Analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/06_Spectral_Mixture_Analysis.ipynb)
+6 [Spectral Mixture Analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/06_Spectral_Mixture_Analysis.ipynb)
 
-7 [Principal Components Analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/07_Principal_Components_Analysis.ipynb)
+7 [Principal Components Analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/07_Principal_Components_Analysis.ipynb)
 
-8 [Tasseled-Cap Transformation.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/08_Tasseled_Cap_Transformation.ipynb)
+8 [Tasseled-Cap Transformation.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/08_Tasseled_Cap_Transformation.ipynb)
 
-9 [Linear trend analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/09_Linear_trend_analysis.ipynb)
+9 [Linear trend analysis.ipynb](https://github.com/yotarazona/scikit-eo/blob/main/examples/notebooks/09_Linear_trend_analysis.ipynb)
 
[…]
