### 2024-06-22 2f234ed
Update atmosCorr.md
