### 2024-06-22 e684728
Update deeplearning.md
