### 2024-06-22 ae87b02
Update linearTrend.md
