### 2024-06-22 d1d2876
Update plot.md
