### 2024-06-24 yotarazona/scikit-eo#18 @KBodolai
fantastic, thanks!
