### 2024-06-25 4aa4360
Update fusionrs.md
