### 2024-06-25 26aeb6e
Update plot.md
