### 2024-06-25 34fed7c
Update writeRaster.md
