### 2024-06-25 yotarazona/scikit-eo#14 @yotarazona
Hi @KBodolai, sorry. It was fixed.
