### 2024-06-25 yotarazona/scikit-eo#14 @KBodolai
I can confirm, great to have this working :)
