### 2024-06-27 8dfb3b7
Re-phrasing summary for more clarity
