### 2024-06-27 yotarazona/scikit-eo#21 @mfbenitezp
Thanks again for your meaningful comments and suggestions; I have addressed the most recent suggestions and left some comments addressing some of the questions included.

1. The r in front of kmeans in rkmeans stands for remote, it is just a name we provide to avoid confusion with other functions.
2. We added the clarification, and yes, we plan to keep extending the package once we get the final feedback from users.
3. We have adjusted the initial paragraph for clarity.
4. We have added the suggested packages
