### 2024-06-27 yotarazona/scikit-eo#21 @yotarazona
Thank you very much @KBodolai  for all your comments. Most of them were addressed, now I will address some other missing issues.

**Scalability:**
Indeed, this is a very important topic but it was already clarified in issue #19 . Please visit this discussion for more details.

**Atmospheric correction**:
This topic is tough because for now only Landsat images can be atmospherically corrected. Honestly, I don't think I'll develop corrections for other images because it's a little outside my area of ​​research and knowledge. The implemented atmospheric correction method is the simplest image-based method but widely used in the community.

**Future versions:**
The idea we have of this package is to try to insert all the preprocessing and processing tools for satellite images and make the analyzes more user-friendly. All the tools presented here were developed out of need for this user, and I still have some processes that I plan to implement in this package. However, other developers/researchers will be invited to implement other new tools to process data not only optical, but radar, for example. Then, God willing, we will see what the future brings in the medium or long term.

Thank you very much again for your comments/suggestions that improved the ```scikit-eo``` package.

Best,
Yonatan
