### 2024-07-01 #6692 @yotarazona
Hi @mikemahoney218 and @KBodolai. We have sent responses to the issues and suggestions made by @KBodolai . We are grateful for these suggestions and comments that improved the package. We look forward to hearing from you if you have any other questions.

Again, thank you very much for your time during this entire review process!. :)
