### 2024-07-02 yotarazona/scikit-eo#21 @KBodolai
Thank you both for addressing this thoroughly!
