### 2024-07-02 #6692 @KBodolai
Hi everyone,

I'm happy with all the edits and responses by @yotarazona , thank you very much for taking the time and for your thoughtful responses!

Everything is checked and all issues are closed, I believe, so I'm happy for the process to continue .
