### 2024-07-02 #6692 @mikemahoney218
Awesome! Thank you so much @dbuscombe-usgs and @KBodolai for your thoughtful reviews here.

@yotarazona , at this point could you please complete the checklist above (making sure to link your archive and mention your final version number here)?

As a bit of housekeeping, I'll be offline the rest of this week so it may be a few days before I can finalize this review.
