### 2024-07-08 5e77ba2
adding details -> paper
