### 2024-07-08 #6692 @yotarazona
Hi @mikemahoney218, some checks below:

1. **Read the text of the paper and offer comments/corrections (as either a list or a pull request)**.
    In the new version, some comments to clarify some ideas were added. (DONE)
2. **Check that the archive title, author list, version tag, and the license are correct**.
    The new version of the paper is within v0.2.28 (https://github.com/yotarazona/scikit-eo/tree/main/paper). License as *Apache Software License 2.0* (DONE)

Let me know if this helps you.
