### 2024-07-08 #6692 @mikemahoney218
Sorry @yotarazona  -- I'm asking you to do the top checklist ("Additional Author Tasks After Review is Complete") -- I'll take care of the editor checklist once you've finished that (including giving me the final version ID, your archive link, etc)
