### 2024-07-08 #6692 @yotarazona
Sorry @mikemahoney218, a little mistake of mine. Checklist below:

  DONE
  The new version with the latest changes: 0.2.31(https://github.com/yotarazona/scikit-eo)
  Zenodo DOI: https://doi.org/10.5281/zenodo.12688708
   DONE
   DONE

Please, let me know if there is anything else missing.
