### 2024-07-09 #6692 @mikemahoney218
Running through my checklist now:

@yotarazona , can you explain why the Zenodo archive doesn't have the same list of authors as your paper? Normally we expect that both the software archive and software paper should share an author list. Is this intentional?

The license of the Zenodo archive should be the same as the software itself (APL-2.0). The current license, CC-BY-4.0, is not an open source license and we cannot accept it.

Below are my comments on the paper. The top section needs to get fixed before I could recommend acceptance, the second section are suggestions I believe would improve the paper but aren't blockers.

## Requested changes

Fig 1: the figure in the published paper is going to look just like the one in the draft. Can you provide a higher resolution figure, potentially with larger text?

Citations:
* Chaves 2020: needs brackets around `{Landsat}` and `{Sentinel}` to preserve capitalization
* "Community" 2019: needs brackets around `{The Turing Way Community}` to preserve the name
* Huss 2018: the journal name is Nature _Climate_ Change, need the full name in your bibtex
* Pott 2021: Need brackets to preserve place name capitalization
* Tarazona 2021: Need brackets to preserve Landsat
* Tarazona 2020: Need brackets to preserve place name capitalization
* Trinder 2020: Can you remove that weird URL? If we have the DOI, we don't need other links.
* Yang 2013: looks like there's too much information in the journal field? Should just be Nature Climate Change

## Other suggestions

Line 18: Consider breaking this into two sentences ("Applications that integrate... impact on urbanization levels. A particular area of interest is measuring and monitoring climate change impacts, including..."). As it is, this one sentence takes 7 lines and uses 5 commas, which makes it a bit hard to follow.

Line 67: I'm pretty confused by this sentence. Consider rephrasing -- maybe breaking it into two sentences, for instance "Its comprehensive features make it well-suited for various applications, from university teaching to cutting-edge research using the most recent machine learning and deep learning techniques. Whether the users are students seeking insights from a satellite image analysis or an experienced researcher looking for advanced tools, scikit-eo offers a valuable resource to support the most valuable methods for environmental studies."

Line 114 and 115: consider "classroom" instead of "lecture room".

Thanks!
