+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Tarazona2021,
-  title={Fusing Landsat and SAR Data for Mapping Tropical Deforestation through Machine Learning Classification and the PVts-β Non-Seasonal Detection Approach},
+  title={Fusing {Landsat} and SAR Data for Mapping Tropical Deforestation through Machine Learning Classification and the PVts-β Non-Seasonal Detection Approach},
   author={Tarazona, Yonatan and Alaitz, Zabala and Xavier, Pons and Antoni, Broquetas and Jakub, Nowosad and Hamdi A., Zurqani},
   journal={Canadian Journal of Remote Sensing},
   volume={47},
@@ -10,7 +10,7 @@
 }
 
 @article{TARAZONA2020100337,
-title = {Monitoring tropical forest degradation using remote sensing. Challenges and opportunities in the Madre de Dios region, Peru},
+title = {Monitoring tropical forest degradation using remote sensing. Challenges and opportunities in the {Madre de Dios} region, {Peru}},
 journal = {Remote Sensing Applications: Society and Environment},
 volume = {19},
 pages = {100337},
@@ -52,7 +52,7 @@ abstract = {The remote sensing science and application communities have develope
 }
 
 @article{POTT2021196,
-title = {Satellite-based data fusion crop type classification and mapping in Rio Grande do Sul, Brazil},
+title = {Satellite-based data fusion crop type classification and mapping in {Rio Grande do Sul, Brazil}},
 journal = {ISPRS Journal of Photogrammetry and Remote Sensing},
 volume = {176},
 pages = {196-210},
@@ -67,7 +67,7 @@ abstract = {Field-scale crop monitoring is essential for agricultural management
 
 @Article{Chaves2020,
 AUTHOR = {Michel E. D. Chaves and Michelle C. A. Picoli and Ieda D. Sanches},
-TITLE = {Recent Applications of Landsat 8/OLI and Sentinel-2/MSI for Land Use and Land Cover Mapping: A Systematic Review},
+TITLE = {Recent Applications of {Landsat} 8/OLI and {Sentinel}-2/MSI for Land Use and Land Cover Mapping: A Systematic Review},
 JOURNAL = {Remote Sensing},
 VOLUME = {12},
 YEAR = {2020},
@@ -150,7 +150,7 @@ DOI = {10.3390/rs12183062}
    author = {The Turing Way Community and Becky Arnold and Louise Bowler and Sarah Gibson and Patricia Herterich and Rosie Higman and Anna Krystalli and Alexander Morley and Martin O'Reilly and Kirstie Whitaker},
    doi = {10.5281/ZENODO.3233986},
    month = {3},
-   title = {The Turing Way: A Handbook for Reproducible Data Science},
+   title = {{The Turing Way}: A Handbook for Reproducible Data Science},
    url = {https://zenodo.org/record/3233986},
    year = {2019},
 }
@@ -158,7 +158,7 @@ DOI = {10.3390/rs12183062}
 @article{Huss2018,
    author = {Matthias Huss and Regine Hock},
    doi = {10.1038/s41558-017-0049-x},
-   journal = {Nature Clim Change},
+   journal = {Nature Climate Change},
    title = {Global-scale hydrological response to future glacier mass loss},
    pages = {135–140},
    volume = {8},
