+++ paper/paper.md
@@ -43,7 +43,7 @@ bibliography: paper.bib
 
 In recent years, a growing body of space-borne and drone imagery has become available with increasing spatial and temporal resolutions. This remotely sensed data has enabled researchers to address and tackle a broader range of challenges effectively by using novel tools and data. However, analysts spend an important amount of time finding the adequate libraries to read and process remotely sensed data.
 
-With an increasing amount of open access data, there is a growing need to account for effective open source tools to read, process and execute analysis that contributes to underpin patterns, changes and trends that are critical for environmental studies. Applications that integrate spatial-temporal data are used to study a variety of complex environmental processes, such as monitoring and assessment of land cover changes [@Chaves2020], crop classifications [@POTT2021196], deforestation [@TARAZONA2018367], impact on urbanization level [@Trinder2020], climate change impacts [@Yang2013]  including assessments of glacier retreat [@Hugonnet2021] and related hydrological change [@Huss2018], biodiversity conservation [@Jeannine2022], and disaster management [@Kucharczyk2021].
+With an increasing amount of open access data, there is a growing need to account for effective open source tools to read, process and execute analysis that contributes to underpin patterns, changes and trends that are critical for environmental studies. Applications that integrate spatial-temporal data are used to study a variety of complex environmental processes, such as monitoring and assessment of land cover changes [@Chaves2020], crop classifications [@POTT2021196], deforestation [@TARAZONA2018367], impact on urbanization level [@Trinder2020] and climate change impacts [@Yang2013]. Other complex environmental processes that are monitored by integrating spatial-temporal data are assessments of glacier retreat [@Hugonnet2021], related hydrological change [@Huss2018], biodiversity conservation [@Jeannine2022] and disaster management [@Kucharczyk2021].
 
 To bridge the gaps in remotely sensed data processing tools, we here introduce **scikit-eo**, a brand-new Python package for satellite remote sensing analysis. Unlike other tools, it is a centralized, scalable, and open-source toolkit due to its flexibility in being adapted into large dataset processing pipelines. It provides central access to the most commonly used Python functions in remote sensing analysis for environmental studies, including machine learning methods. **scikit-eo** stands out with its ability to be used in various settings, from a lecturer room to a crucial part of any Python environment in a research project. The majority of the tools included in **scikit-eo** are derived from peer-reviewed scientific publications, ensuring their reliability and accuracy.
 
@@ -61,7 +61,9 @@ In this paper, we introduce **scikit-eo**, a specialized library with tailored a
 
 # Audience
 
-**scikit-eo** is an adaptable Python package that covers multiple users, including students, remote sensing professionals, environmental analysis researchers, and organizations looking for satellite image analysis. Its comprehensive features make it well-suited for various applications, such as university teaching, that includes technical and practical sessions and cutting-edge research using the most recent machine learning and deep learning techniques applied to remote sensing whether the users are students seeking insights from a satellite image analysis or an experienced researcher looking for advanced tools, **scikit-eo** offers a valuable resource to support the most valuable methods for environmental studies.
+**scikit-eo** is an adaptable Python package that covers multiple users, including students, remote sensing professionals, environmental analysis researchers, and organizations looking for satellite image analysis. Its tools and algorithms implemented make it well-suited for various applications, such as university teaching, that includes technical and practical sessions and cutting-edge research using the most recent machine learning and deep learning techniques applied to remote sensing. 
+
+This python package provides key tools for both students seeking insights from a satellite image analysis or an experienced researcher looking for advanced tools. **scikit-eo** offers a valuable resource to support the most valuable methods for environmental studies.
 
 ### **scikit-eo** as a research tool:
 
[…]
