### 2024-07-09 #6692 @yotarazona
Hi @mikemahoney218, I am new to zenodo :).
I had to edit the author list to add the other authors and modify the license as well. Now you can verify that both the authors and the license were fixed: https://doi.org/10.5281/zenodo.12688708.

**Requested changes:**
DONE. Thank you very much for these suggestions.
**Other suggestions:**
DONE. Thank you again for these suggestions. I agree and I think they improve the flow of the reading. You can see the changes in the updated paper.md https://github.com/yotarazona/scikit-eo/blob/main/paper/paper.md.

About the Fig. 1, we have two higher resolution figures in the paper folder: https://github.com/yotarazona/scikit-eo/tree/main/paper.

Let me know if you have any other comments.
