### 2024-07-09 #6692 @mikemahoney218
@yotarazona , looking closer at Figure 1, I think the main issue is that the figure is too wide for the paper format -- the aspect ratio is so wide (and relatively short), so when it gets automatically resized for the paper, the text gets a bit squashed and winds up looking a bit fuzzy and harder to read. If you wanted it to be more legible, I wonder if you'd be able to "rotate" the figure so that it's taller than it is wide -- ideally by changing the orientation of the flow chart, but also potentially by rotating the image in Photoshop or similar.

If you don't mind how Figure 1 looks, I suppose this doesn't need to be a blocker -- but I think it would improve the legibility for the final paper if the figure is a bit clearer!
