### 2024-07-09 586450e
Fix journal name
