+++ paper/paper.bib
@@ -102,7 +102,7 @@ DOI = {10.3390/rs12183062}
    doi = {10.1038/nclimate1908},
    issn = {1758-6798},
    issue = {10},
-   journal = {Nature Climate Change 2013 3:10},
+   journal = {Nature Climate Change},
    keywords = {Attribution,Climate,change impacts},
    month = {9},
    pages = {875-883},
