### 2024-07-09 #6692 @mikemahoney218
Also: opened https://github.com/yotarazona/scikit-eo/pull/24, https://github.com/yotarazona/scikit-eo/pull/25, https://github.com/yotarazona/scikit-eo/pull/26 to fix bibtex issues
