### 2024-07-09 #6692 @yotarazona
Hi @mikemahoney218, a new workflow was uploaded in the paper folder: https://github.com/yotarazona/scikit-eo/blob/main/paper/workflow_updated.png.

Open issues were merged as well. Thanks for that!
