### 2024-07-09 #6692 @mikemahoney218
@yotarazona I see the new figure is uploaded -- would you be able to change `paper.md` so that it uses the new figure?
