+++ paper/paper.md
@@ -57,7 +57,7 @@ In this paper, we introduce **scikit-eo**, a specialized library with tailored a
 
 **scikit-eo** is an open-source package built entirely in Python through Object-Oriented Programming and Structured Programming that provides a helpful variety of remote sensing tools (see \autoref{fig:workflow}), from primary and exploratory functions to more advanced methods to classify, calibrate, or fuse satellite imagery. Depending on the users' needs, **scikit-eo** can provide the basic but essential land cover characterization mapping, including the confusion matrix and the required metrics such as user's accuracy, producer's accuracy, omission and commission errors. These required metrics can be combined as a pandas ```DataFrame``` object. Furthermore, a class prediction map is a result of land cover mapping, i.e., a land cover map, which represents the output of the classification algorithm or the output of the segmentation algorithm. These two outcomes must include uncertainties with a confidence level (e.g., at $95$% or $90$%). All required metrics from the confusion matrix can be easily computed and included confidence levels with **scikit-eo** following guidance proposed by @OLOFSSON201442. Other useful tools for remote sensing analysis can be found in this package; and for more information about the full list of the supported functions, tutorials as well as how to install and execute the package within a Python setting, visit the [scikit-eo](https://yotarazona.github.io/scikit-eo/) website.
 
-![Workflow of main functionalities of the *scikit-eo* python package as well as outputs that can be obtained by using the tools developed. \label{fig:workflow}](workflow_basic.png){ width=100% }
+![Workflow of main functionalities of the *scikit-eo* python package as well as outputs that can be obtained by using the tools developed. \label{fig:workflow}](workflow_updated.png){ width=100% }
 
 # Audience
 
