### 2024-07-09 #6692 @mikemahoney218
(And as one brief note: Kristen, our EiC, is at SciPy this week, so there may be a bit of a delay before she's able to process this submission -- just so you know!)
