### 2024-07-09 #6692 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
