### 2024-07-16 #6692 @kthyng
Congratulations on your new publication @ytarazona! Many thanks to @mikemahoney218 and to reviewers @dbuscombe-usgs and @KBodolai for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
