### 2024-07-16 #6692 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
