### 2024-07-16 #6692 @kthyng
@yotarazona If you're interested in joining as a reviewer, please register here: https://reviewers.joss.theoj.org/!
