### 2024-07-16 #6692 @yotarazona
> Congratulations on your new publication @ytarazona! Many thanks to @mikemahoney218 and to reviewers @dbuscombe-usgs and @KBodolai for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.

Thank you so much @kthyng, @mikemahoney218, @dbuscombe-usgs and @KBodolai for your time and effort during this process!. Hard work for everyone!. Thanks a lot.
A hug!.
